credential-management/fedcm-logout.sub.https.html

<!DOCTYPE html>
<title>Federated Credential Management logout() API.</title>
<link rel="help" href="https://wicg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script type="module">
  import {set_fedcm_cookie} from './support/fedcm-helper.js';

  const url_prefix = 'https://{{host}}:{{ports[https][0]}}/credential-management/support/';
  const test_options = {
    federated: {
      providers: [{
        url: url_prefix,
        clientId: '1',
        // hint has to match the account id in support/accounts.json
        hint: '1234',
      }]
    }
  };
  const test_options_no_hint = {
    federated: {
      providers: [{
        url: url_prefix,
        clientId: '1',
        nonce: '2',
      }]
    }
  };

  promise_test(async t => {
    await set_fedcm_cookie();
    const cred = await navigator.credentials.get(test_options);
    assert_equals(cred.idToken, "token");

    await cred.logout();
  }, "logout should not throw an exception.");

  promise_test(async t => {
    await set_fedcm_cookie();
    const cred = await navigator.credentials.get(test_options_no_hint);
    assert_equals(cred.idToken, "token");

    return promise_rejects_dom(t, "InvalidStateError", cred.logout());
  }, "logout should require a hint.");
</script>
